Extend Mapping submap_floor and submap_residue arrays.
The submap arrays in Mapping are size 15, but should be size 16.
The number of submap entries is in the range [1,16] and the mux
values that these arrays are indexed by are in the range [0,15],
as long as they are less than the submap count. Per the Vorbis
spec, these ranges are correct, and nothing indicates that
submap 15 should be considered invalid:

https://xiph.org/vorbis/doc/Vorbis_I_spec.html#x1-650004.2.4

Also moves Mapping::chan to the front of Mapping to reduce the
padding overhead of this struct, which was expanded by extending
the submap arrays.

https://github.com/nothings/stb/pull/1312

This is GHSL-2023-168/CVE-2023-45678 -- see:
https://github.com/nothings/stb/pull/1556